fix: Cloudflare Pages Functionsに移行してSSRアダプターを削除
@astrojs/cloudflare v13のworkerdビルドエラーを回避するため、
APIルートをAstro SSRからCloudflare Pages Functionsに移行した。

- astro.config.mjs: output='static'に戻しadapterを削除
- functions/api/: Pages Functionsとしてauth・quizAPIを再実装
- src/pages/api/: Astro APIルートを削除
- src/pages/history.astro: クライアントサイドレンダリングに変更
- src/env.d.ts: App.Localsの拡張を削除（static不要）
- 各ページのexport const prerender=trueを削除（static時のデフォルト）
