+++ astro.config.mjs
@@ -1,12 +1,10 @@
 // @ts-check
 import { defineConfig } from 'astro/config';
 import tailwindcss from '@tailwindcss/vite';
-import cloudflare from '@astrojs/cloudflare';
 
 // https://astro.build/config
 export default defineConfig({
-  output: 'server',
-  adapter: cloudflare({ prerenderEnvironment: 'node' }),
+  output: 'static',
   vite: {
     plugins: [tailwindcss()],
   },
