+++ functions/api/auth/logout.ts
@@ -1,0 +1,8 @@
+import { clearSessionCookie } from '../../../src/lib/session';
+
+export const onRequestPost: PagesFunction<Env> = () => {
+  return new Response(null, {
+    status: 302,
+    headers: { Location: '/', 'Set-Cookie': clearSessionCookie() },
+  });
+};
