+++ src/env.d.ts
@@ -7,9 +7,3 @@ interface Env {
   GITHUB_CLIENT_SECRET: string;
   JWT_SECRET: string;
 }
-
-type Runtime = import('@astrojs/cloudflare').Runtime<Env>;
-
-declare namespace App {
-  interface Locals extends Runtime {}
-}
