+++ src/pages/index.astro
@@ -1,5 +1,4 @@
 ---
-export const prerender = true;
 import BaseLayout from '../layouts/BaseLayout.astro';
 import { curriculum } from '../data/curriculum';
 import { questions } from '../data/questions';
