+++ src/pages/quiz.astro
@@ -1,5 +1,4 @@
 ---
-export const prerender = true;
 import BaseLayout from '../layouts/BaseLayout.astro';
 import { questions } from '../data/questions';
